+++ keys/types.go
@@ -1,11 +1,22 @@
 package keys
 
 import (
+	"fmt"
 	"sort"
 
 	crypto "github.com/tendermint/go-crypto"
+	wire "github.com/tendermint/go-wire"
 	data "github.com/tendermint/go-wire/data"
 )
+
+// Storage has many implementation, based on security and sharing requirements
+// like disk-backed, mem-backed, vault, db, etc.
+type Storage interface {
+	Put(name string, salt []byte, key []byte, info Info) error
+	Get(name string) (salt []byte, key []byte, info Info, err error)
+	List() (Infos, error)
+	Delete(name string) error
+}
 
 // Info is the public information about a key
 type Info struct {
@@ -71,4 +82,53 @@ type Manager interface {
 	Get(name string) (Info, error)
 	Update(name, oldpass, newpass string) error
 	Delete(name, passphrase string) error
+}
+
+/**** MockSignable allows us to view data ***/
+
+// MockSignable lets us wrap arbitrary data with a go-crypto signature
+type MockSignable struct {
+	Data      []byte
+	PubKey    crypto.PubKey
+	Signature crypto.Signature
+}
+
+var _ Signable = &MockSignable{}
+
+// NewMockSignable sets the data to sign
+func NewMockSignable(data []byte) *MockSignable {
+	return &MockSignable{Data: data}
+}
+
+// TxBytes returns the full data with signatures
+func (s *MockSignable) TxBytes() ([]byte, error) {
+	return wire.BinaryBytes(s), nil
+}
+
+// SignBytes returns the original data passed into `NewSig`
+func (s *MockSignable) SignBytes() []byte {
+	return s.Data
+}
+
+// Sign will add a signature and pubkey.
+//
+// Depending on the Signable, one may be able to call this multiple times for multisig
+// Returns error if called with invalid data or too many times
+func (s *MockSignable) Sign(pubkey crypto.PubKey, sig crypto.Signature) error {
+	s.PubKey = pubkey
+	s.Signature = sig
+	return nil
+}
+
+// Signers will return the public key(s) that signed if the signature
+// is valid, or an error if there is any issue with the signature,
+// including if there are no signatures
+func (s *MockSignable) Signers() ([]crypto.PubKey, error) {
+	if s.PubKey.Empty() {
+		return nil, fmt.Errorf("no signers")
+	}
+	if !s.PubKey.VerifyBytes(s.SignBytes(), s.Signature) {
+		return nil, fmt.Errorf("invalid signature")
+	}
+	return []crypto.PubKey{s.PubKey}, nil
 }